rename all as scripts.package_info
