+++ docs/src/apiref.rst
@@ -79,6 +79,7 @@ Modules:
     topic_coherence/probability_estimation
     topic_coherence/segmentation
     topic_coherence/text_analysis
+    scripts/package_info
     scripts/glove2word2vec
     scripts/make_wikicorpus
     scripts/word2vec_standalone
@@ -99,4 +100,3 @@ Modules:
     summarization/syntactic_unit
     summarization/textcleaner
     viz/poincare
-    gensim.rst
