--- docs/src/gensim.rst
@@ -1,9 +1,0 @@
-:mod:`gensim` -- CLI interface of Gensim
-========================================
-
-.. automodule:: gensim
-    :synopsis: CLI interface of Gensim.
-    :members:
-    :inherited-members:
-    :undoc-members:
-    :show-inheritance:
