+++ docs/src/scripts/package_info.rst
@@ -1,0 +1,9 @@
+:mod:`scripts.package_info` -- Information about gensim package
+===============================================================
+
+.. automodule:: gensim.scripts.package_info
+    :synopsis: Information about gensim package.
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
