--- gensim/__main__.py
@@ -1,5 +1,0 @@
-from gensim import main
-
-
-if __name__ == "__main__":
-    main()
